% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{checkSavedAccessToken}
\alias{checkSavedAccessToken}
\title{Check the token saved by saveTokenToFile}
\usage{
checkSavedAccessToken(savedFile = "~/.access_token")
}
\arguments{
\item{savedFile}{The file read that contains the token.}
}
\value{
invisibly returns a list of information about the token
}
\description{
Reads a token out of savedFile.  If there is a valid token, the email of the
service account is displayed.  Otherwise, there is an error, you should consider
recreating the token using \code{\link[=saveTokenToFile]{saveTokenToFile()}}
}
\seealso{
\code{\link[=saveTokenToFile]{saveTokenToFile()}}
}
