% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{cloud_run_qaqc}
\alias{cloud_run_qaqc}
\title{Run QA/QC for Connect for Cancer Prevention}
\usage{
cloud_run_qaqc(service_account_key_path)
}
\arguments{
\item{service_account_key_path}{The service account key file.}
}
\value{
a list holding the output of the QAQC code
}
\description{
Calls the QA/QC API on Google cloud platform.  Currently the code requires a service
account key for prod/stage environment.  In the future, code will run with the user's
credentials obtained via OAUTH2.
}
\details{
Note: this takes a few minutes to run.
}
