% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectAnalystR-package.R
\docType{package}
\name{ConnectAnalystR-package}
\alias{ConnectAnalystR}
\alias{ConnectAnalystR-package}
\title{ConnectAnalystR: R Functions for NCI/DCEG Connect Analysts}
\description{
This packages provides a set of functions that a DCEG connect analyst might find useful.
}
\author{
\strong{Maintainer}: Russ Daniel \email{daniel.russ@nih.gov} (\href{https://orcid.org/0000-0003-4040-4416}{ORCID})

}
\keyword{internal}
