% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectBQ.R
\name{getIndustryData}
\alias{getIndustryData}
\title{getIndustryData}
\usage{
getIndustryData(project = preferences$project, env = preferences$env)
}
\arguments{
\item{project}{\itemize{
\item The project which is billed for the BQ processing.
}}

\item{env}{\itemize{
\item Either  "dev", "stage", or "prod"
}}
}
\value{
\itemize{
\item tibble with industry data:
The Connect_ID, a column of integer64's representingthe Connect_ID;
CurrentIndustryCID, a list column containing 1 or more ConceptIDs;
CurrentIndustry, a list column containing 1 or more strings decyphered from the concept id;
CurrentIndustryOther, a column of characters that represents what a participant put in the
other field of the the question.
LongestIndustryCID,LongestIndustry,LongestIndustryOther are the same as the CurrentXX except
they are for the Longest Job question instead of the current job question.
}
}
\description{
getIndustryData
}
