% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectBQ.R
\name{getOccupationData}
\alias{getOccupationData}
\title{Get Occupation data from the Connect Data.}
\usage{
getOccupationData(project = preferences$project, env = preferences$env)
}
\arguments{
\item{project}{\itemize{
\item The project which is billed for the BQ processing.
}}

\item{env}{\itemize{
\item Either  "dev", "stage", or "prod"
}}
}
\value{
tibble with data.
}
\description{
Calls the BQ API.  The user will have to go through the OAuth dance.  To make sure you
have access to the data, you will want to run bq_auth().
}
