% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectBQ.R
\name{makeOccupationBarChart}
\alias{makeOccupationBarChart}
\title{Makes a bar chart for the Occupational Data Downloaded from the Connect Data}
\usage{
makeOccupationBarChart(occData)
}
\arguments{
\item{occData}{The occupational data returned from  \code{\link{getOccupationData}}.}
}
\value{

}
\description{
Makes a bar chart that from the Occupational Data in Connect.
}
