% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectBQ.R
\name{pivotOccupationData}
\alias{pivotOccupationData}
\title{Pivot the Occupational data}
\usage{
pivotOccupationData(data, ...)
}
\arguments{
\item{data}{Occupational data from getOccupationData}

\item{...}{alternatively, pass in the variables for getOccupationalData and this function
will call getOccupationalData and pivot in one step}
}
\value{
a tibble with Connect Id, longest/Current job, JobTitle, Selection, and rank (1-4,5+)
}
\description{
Pivot the Occupational data
}
