% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{saveTokenToFile}
\alias{saveTokenToFile}
\title{saveTokenToFile}
\usage{
saveTokenToFile(jsonFile, saveTo = "~/.access_token")
}
\arguments{
\item{jsonFile}{Path to a service account key JSON file}

\item{saveTo}{Path of the output text file containing the service account
token}
}
\description{
Gets a service account token and writes it to a plain text file.
}
