% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_project}
\alias{set_project}
\alias{set_env}
\title{set the preferences/default parameters}
\usage{
set_project(project)

set_env(env = c("dev", "stage", "prod"))
}
\arguments{
\item{project}{The GCP project getting billed}

\item{env}{prod, stage, or dev Connect data}
}
\description{
\code{set_env} sets the environment for the data. It can from from the development (dev),
staging (stage) or production (prod) Connect environments.

\verb{set project} sets the project that will get billed for the data processing and download.
}
