% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectBQ.R
\name{makeDetailedOccupationBarChart}
\alias{makeDetailedOccupationBarChart}
\alias{makeOccupationBarChart}
\title{Makes a bar chart for the Occupational Data Downloaded from the Connect Data}
\usage{
makeDetailedOccupationBarChart(pivottedData, ...)

makeOccupationBarChart(occData)
}
\arguments{
\item{pivottedData}{Occupational data returned from \code{\link{pivotOccupationData}}.}

\item{...}{Parameters passed to getOccupationData(), must include both and project,env}

\item{occData}{The occupational data returned from  \code{\link{getOccupationData}}.}
}
\value{
ggplot object of barplot ready for displaying or saving with ggsave.
}
\description{
Makes a bar chart that from the Occupational Data in Connect.

makeOccupationBarChart() returns a bar chart displaying whether or not a job was
coded to the top 4.

makeDetailedOccupationBarChart() returns a bar chart showing the count of the rank for
all the Connect data.  The count is slight lower in this plot because the first 324
jobs did not have the rank.
}
